scrapy/real-estate-crawl/realestatevn/requirements.txt
scrapy >= 1.8.0
scrapy-user-agents >= 0.1.1
scrapy-rotating-proxies >= 0.6.2
scrapy-random-useragent >= 0.2
scrapy-splash >= 0.7.2
pyyaml